remove useless files & refactor webpack config
