+++ package.json
@@ -4,8 +4,8 @@
   "main": "index.js",
   "scripts": {
     "dev": "no dev",
-    "dev-server": " NODE_ENV=development nodemon ./server/server.js",
-    "prod": "NODE_ENV=production webpack --progress --colors",
+    "dev-server": "NODE_ENV=development nodemon ./server/server.js",
+    "prod": "no prod",
     "prod-server": "NODE_ENV=production node ./server/server.js",
     "lint": "no lint"
   },
@@ -55,6 +55,7 @@
     "ejs": "^2.5.6",
     "express": "^4.15.3",
     "lodash": "^4.17.4",
+    "normalize.css": "^7.0.0",
     "react": "^15.6.1",
     "react-dom": "^15.6.1",
     "react-fa": "^4.2.0",
