+++ server/views/index.ejs
@@ -8,7 +8,6 @@
   </head>
   <body>
     <div id="app"></div>
-    <script src="/static/vendor.bundle.js"></script>
     <script src="/static/app.js"></script>
   </body>
 </html>
