--- src/actions/indexActions.js
@@ -1,7 +1,0 @@
-export function getIndex(){
-    return ((dispatch) => {
-        dispatch({
-            type: "DEFAULT_STATE"
-        })
-    })
-}
