--- src/constants/index.json
@@ -1,3 +1,0 @@
-{
-    "home": "Hello world"
-}
