+++ src/index.js
@@ -1,0 +1,1 @@
+console.log('home index!');
