--- src/reducers/index.js
@@ -1,7 +1,0 @@
-import { combineReducers } from 'redux'
-
-import IndexReducer from './indexReducer.js'
-
-export default combineReducers({
-    index: IndexReducer
-})
