+++ src/styles/index.css
@@ -1,0 +1,1 @@
+@import '~normalize.css/normalize.css'
