--- src/styles/style.js
@@ -1,7 +1,0 @@
-/**
- * 
- * All all '.less' files here to be compiled into single css
- * 
-**/
-
-require("./index.less")
